Move scahding from topomodule to main thread, so it is available for all modules
