+++ sandbox/projector/cmap.py
@@ -1,26 +1,21 @@
 import matplotlib.pyplot as plt
-import matplotlib
+import matplotlib.colors as mcolors
 import copy
 import panel as pn
-pn.extension()
 import weakref
+pn.extension()
 
 
 class CmapModule:
     """
     Class to manage changes in the colormap and plot in the desired projector figure
     """
-    def __init__(self, cmap='gist_earth', over=None, under=None,
-                 bad=None, norm=None, lot=None, vmin=None, vmax=None,
-                 extent=None):
+    def __init__(self, cmap='gist_earth', norm=None, lot=None, vmin=None, vmax=None, extent=None):
         """
         Initialize the colormap to plot using imshow()
         Args:
             cmap (str or plt.Colormap): Matplotlib colormap, given as name or instance.
-            over (e.g. str): Color used for values above the expected data range.
-            under (e.g. str): Color used for values below the expected data range.
-            bad (e.g. str): Color used for invalid or masked data values.
-            norm: Future feature!
+            norm: Apply norm to imshow
             lot: Future feature!
             vmin (float): ...
             vmax (float): ...
@@ -42,22 +37,24 @@ class CmapModule:
         self.cmap = plt.cm.get_cmap(cmap)#self.set_cmap(plt.cm.get_cmap(cmap), over, under, bad)
         self._cmap = None
         self.norm = norm
-        self.lot = lot  # TODO: Future feature
+        self.lot = lot  # TODO: Future feature ??
         self.col = None
-        self._col = None #weakreference of self.col
+        self._col = None  # weakreference of self.col
         self.active = True
 
+        # Relief shading
+        self.relief_shading = True
+        self.shade = None
+        self._shade = None  # weakreference of self.shade
+
+        self.azdeg = 315
+        self.altdeg = 45
+        self.ve = 0.25
+
     def update(self, sb_params: dict):# data, extent, ax, cmap, norm):
-        """
-        if self.active:
-            self.set_data(data)
-            self.set_cmap(cmap, 'k', 'k', 'k')
-        else:
-            self.delete_image() #Todo
-        """
         active = sb_params.get('active_cmap')
+        active_shade = sb_params.get('active_shading')
         ax = sb_params.get('ax')
-        #ax.texts = [] #TODO: if this is not cleared then the labels breaks the thread
         data = sb_params.get('frame')
         cmap = sb_params.get('cmap')
         norm = sb_params.get('norm')
@@ -68,8 +65,7 @@ class CmapModule:
         if active and self.active:
             if self._col is not None and self._col() not in ax.images:
                 self.col = None
-            if self.col is None:# and self.active and active:
-            #if len(ax.images)==0:
+            if self.col is None:
                 self.render_frame(data, ax, vmin=self.vmin, vmax=self.vmax, extent=extent[:4])
             else:
                 self.set_data(data)
@@ -78,13 +74,42 @@ class CmapModule:
                 self.set_extent(extent)
                 sb_params['cmap'] = self.cmap
         else:
-            #ax.images[0].remove()
             if self.col is not None:
                 self.col.remove()
                 self.col = None
             if self._col is not None and self._col() in ax.images:
                 ax.images.remove(self._col)
+
+        if active_shade and self.relief_shading:
+            if self._shade is not None and self._shade() not in ax.images:
+                self.shade = None
+            # Note: 180 degrees are subtracted because visualization in Sandbox is upside-down
+            ls = mcolors.LightSource(azdeg=self.azdeg - 180, altdeg=self.altdeg)
+            frame_shade = ls.shade(data, cmap=plt.cm.gray, vert_exag=self.ve, blend_mode='hsv')
+            if self.shade is None:
+                self.render_shade(frame_shade, ax, extent=extent[:4])
+            else:
+                self.shade.set_data(frame_shade)
+                self.shade.set_extent(extent[:4])
+        else:
+            if self.shade is not None:
+                self._delete_shade()
+            if self._shade is not None and self._shade() in ax.images:
+                ax.images.remove(self._shade)
         return sb_params
+
+    def _delete_shade(self):
+        """remove relief shading from the axes if present"""
+        if self.shade:
+            self.shade.remove()
+        self.shade = None
+
+    def render_shade(self, frame_shade, ax, extent=None):
+        """Renders a new image or actualizes the current one"""
+        self.shade = ax.imshow(frame_shade, origin='lower', extent=extent, alpha=0.4, aspect='auto', zorder=-4)
+        self._shade = weakref.ref(self.shade)
+        return None
+
 
     def set_extent(self, extent):
         self.col.set_extent(extent[:4])
@@ -124,16 +149,6 @@ class CmapModule:
         self.col.set_cmap(cmap)
         return None
 
-    #def set_array(self, data):
-    #    """
-    #    Change the numpy array that is being plotted without the need to errase the pcolormesh figure
-    #    Args:
-    #        data:
-    #    Returns:
-    #    """
-    #    self.col.set_array(data.ravel())
-    #    return None
-
     def set_data(self, data):
         """
         Change the numpy array that is being plotted without the need to errase the imshow figure
@@ -141,10 +156,8 @@ class CmapModule:
             data:
         Returns:
         """
-        #masked = numpy.ma.masked_outside(data, self.vmin, self.vmax )
         self.col.set_data(data)
         self.col.set_clim(vmin=self.vmin, vmax=self.vmax)
-        #self.col.set_aspect('auto')
         return None
 
     def render_frame(self, data, ax, vmin=None, vmax=None, extent=None):
@@ -154,16 +167,9 @@ class CmapModule:
         if vmax is None:
             vmax = self.vmax
 
-
-        #self.col = ax.pcolormesh(data, vmin=vmin, vmax=vmax,
-        #                         cmap=self.cmap, norm=self.norm,
-        #                         shading='nearest')
-        #self.col = matplotlib.image.AxesImage(ax, cmap=self.cmap, norm=self.norm,
-                                                #  origin='lower',zorder=-1)
-        #self.set_data(data)
         self.col = ax.imshow(data, vmin=vmin, vmax=vmax,
                              cmap=self.cmap, norm=self.norm,
-                             origin='lower', aspect='auto', zorder=-1, extent=extent)
+                             origin='lower', aspect='auto', zorder=-5, extent=extent)
         self._col = weakref.ref(self.col)
 
         ax.set_axis_off()
@@ -178,12 +184,10 @@ class CmapModule:
 
     def show_widgets(self):
         self._create_widgets()
-
         panel = pn.Column("<b> Colormap </b>",
-                          self._widget_plot_cmap,
                           self._widget_plot_colormap,
-                          )
-
+                          self._widget_plot_cmap,
+                          self._widget_lightsource())
         return panel
 
     def _create_widgets(self):
@@ -208,7 +212,46 @@ class CmapModule:
         self.active = event.new
 
     def _callback_plot_cmap(self, event):
-        #self.lock.acquire()
         self._cmap = plt.cm.get_cmap(event.new)
-        #self.set_cmap(cmap, 'k', 'k','k')
-        #self.lock.release()
+
+    def _widget_lightsource(self):
+        self._widget_relief_shading = pn.widgets.Checkbox(name='Show relief shading',
+                                                          value=self.relief_shading)
+        self._widget_relief_shading.param.watch(self._callback_relief_shading, 'value',
+                                                onlychanged=False)
+        self._widget_azdeg = pn.widgets.FloatSlider(name='Azimuth',
+                                                    value=self.azdeg,
+                                                    start=0.0,
+                                                    end=360.0)
+        self._widget_azdeg.param.watch(self._callback_lightsource_azdeg, 'value')
+
+        self._widget_altdeg = pn.widgets.FloatSlider(name='Altitude',
+                                                     value=self.altdeg,
+                                                     start=0.0,
+                                                     end=90.0)
+        self._widget_altdeg.param.watch(self._callback_lightsource_altdeg, 'value')
+
+        self._widget_ve = pn.widgets.Spinner(name="Vertical Exageration", value=self.ve,
+                                             step=0.01)
+        self._widget_ve.param.watch(self._callback_ve, 'value', onlychanged=False)
+
+        widgets = pn.WidgetBox(self._widget_relief_shading,
+                               self._widget_azdeg,
+                               self._widget_altdeg,
+                               self._widget_ve
+                               )
+
+        panel = pn.Column("<b> Lightsource </b> ", widgets)
+        return panel
+
+    def _callback_relief_shading(self, event):
+        self.relief_shading = event.new
+
+    def _callback_ve(self, event):
+        self.ve = event.new
+
+    def _callback_lightsource_altdeg(self, event):
+        self.altdeg = event.new
+
+    def _callback_lightsource_azdeg(self, event):
+        self.azdeg = event.new
